added logout api endpoint. missing if should add the invalidated token to db or through frontend
