+++ backend/app.js
@@ -1,7 +1,7 @@
 import Fastify from 'fastify';
 import dotenv from 'dotenv';
 import jwtPlugin from './plugins/jwt-plugin.js';
-import loginRoutes from './routes/loginRoutes.js';
+import authRoutes from './routes/authRoutes.js';
 import userRoutes from './routes/userRoutes.js';
 
 // Load environment variables
@@ -12,7 +12,7 @@ const fastify = Fastify({ logger: true });
 
 // Register routes
 fastify.register(jwtPlugin);
-fastify.register(loginRoutes);
+fastify.register(authRoutes);
 fastify.register(userRoutes);
 
 export default fastify;
