+++ backend/test/authTest.http
@@ -1,0 +1,31 @@
+### LOGIN
+POST http://localhost:3001/login
+Content-Type: application/json
+
+{
+	"username": "kmatjuhi",
+	"password": "password"
+}
+
+### INVALID PASSWORD
+POST http://localhost:3001/login
+Content-Type: application/json
+
+{
+	"username": "kmatjuhi",
+	"password": "passwdawdk"
+}
+
+### CREATE USER KIM
+POST http://localhost:3001/register HTTP/1.1
+Content-Type: application/json
+
+{
+	"username": "kim23213",
+	"email": "kmatjuhi@hive.fr",
+	"password": "password"
+}
+
+### LOGOUT
+POST http://localhost:3001/logout HTTP/1.1
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImttYXRqdWhpIiwiaWQiOjEsImlhdCI6MTc0MzY2NTkyOSwiZXhwIjoxNzQzNjY5NTI5fQ.1vscf9CFyrdwH37edwEUOAgT2VUm5JJIQDrAGXOS_JA
