--- backend/test/loginTest.http
@@ -1,17 +1,0 @@
-### LOGIN
-POST http://localhost:3001/login
-Content-Type: application/json
-
-{
-	"username": "kmatjuhi",
-	"password": "password"
-}
-
-### INVALID PASSWORD
-POST http://localhost:3001/login
-Content-Type: application/json
-
-{
-	"username": "kmatjuhi",
-	"password": "passwdawdk"
-}
